Pull request #116: AG-37879 (Start using int32 instead of int64 for filterId's type)
Merge in ADGUARD-CORE-LIBS/filter-list-manager from AG-37879 to master

Squashed commit of the following:

commit fa1373d27ced32122a5f30b332c6f1bb756bbf01
Author: Boris Bukin <boris@adguard.com>
Date:   Wed Nov 20 16:17:10 2024 +0300

    issue #AG-37879, regex to proj 2

commit 48835eb30bec397684e8a974b32914c396a185d9
Author: Boris Bukin <boris@adguard.com>
Date:   Wed Nov 20 16:15:41 2024 +0300

    issue #AG-37879, regex to proj

commit 786978914bd59f91f3bb7bdf5d84d5d775ba2a31
Author: Boris Bukin <boris@adguard.com>
Date:   Wed Nov 20 15:04:49 2024 +0300

    issue #AG-37879, long to int

+++ crates/ffi/src/platforms/windows/AdGuard.FilterListManagerProtobuf/AdGuard.FilterListManagerProtobuf.SampleApp/Program.cs
@@ -30,7 +30,7 @@ namespace AdGuard.FilterListManagerProtobuf.SampleApp
                 flm.PullMetadata();
                 flm.UpdateFilters(false, 0, false);
                 
-                flm.EnableFilterLists(new long[] {1, 2, 255}, true);
+                flm.EnableFilterLists(new int[] {1, 2, 255}, true);
                 FullFilterList customFilter = flm.InstallCustomFilterFromString(
                     string.Empty,
                     1000000000,
@@ -41,8 +41,8 @@ namespace AdGuard.FilterListManagerProtobuf.SampleApp
                     "Desc");
                 bool localeResult = flm.ChangeLocale("ru_RU");
                 flm.LiftUpDatabase();
-                flm.EnableFilterLists(new long[] {1, 2, 255}, true);
-                flm.InstallFilterLists(new long[] {1, 2, 255}, true);
+                flm.EnableFilterLists(new int[] {1, 2, 255}, true);
+                flm.InstallFilterLists(new int[] {1, 2, 255}, true);
 
                 FilterListRules rules1 = new FilterListRules
                 {
@@ -63,7 +63,7 @@ namespace AdGuard.FilterListManagerProtobuf.SampleApp
                 FileUtils.Touch(blobPath);
                 flm.SaveRulesToFileBlob(customFilter.Id, blobPath);
                 flm.GetFullFilterListById(customFilter.Id);
-                flm.ForceUpdateFiltersByIds(new long[] { 1, 2 }, 0);
+                flm.ForceUpdateFiltersByIds(new int[] { 1, 2 }, 0);
                 customFilter = flm.InstallCustomFilterList(
                     "https://filters.adtidy.org/extension/safari/filters/101_optimized.txt",
                     true,
